test error handling in query register
