+++ reductstore/src/ext/ext_repository.rs
@@ -1,6 +1,9 @@
 // Copyright 2025 ReductSoftware UG
 // Licensed under the Business Source License 1.1
 
+mod create;
+mod load;
+
 use crate::asset::asset_manager::ManageStaticAsset;
 use crate::ext::filter::ExtWhenFilter;
 use crate::storage::query::base::QueryOptions;
@@ -9,16 +12,14 @@ use crate::storage::query::QueryRx;
 use async_trait::async_trait;
 use dlopen2::wrapper::{Container, WrapperApi};
 use futures_util::StreamExt;
-use log::{error, info};
 use reduct_base::error::ErrorCode::NoContent;
 use reduct_base::error::ReductError;
 use reduct_base::ext::{
     BoxedCommiter, BoxedProcessor, BoxedReadRecord, BoxedRecordStream, ExtSettings, IoExtension,
 };
 use reduct_base::msg::entry_api::QueryEntry;
-use reduct_base::{internal_server_error, no_content, unprocessable_entity};
+use reduct_base::{no_content, unprocessable_entity};
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::pin::Pin;
 use std::sync::Arc;
 use std::time::Instant;
@@ -82,62 +83,6 @@ struct ExtRepository {
 
     #[allow(dead_code)]
     embedded_extensions: Vec<Box<dyn ManageStaticAsset + Sync + Send>>, // we need to keep them from being cleaned up
-}
-
-impl ExtRepository {
-    fn try_load(
-        paths: Vec<PathBuf>,
-        embedded_extensions: Vec<Box<dyn ManageStaticAsset + Sync + Send>>,
-        settings: ExtSettings,
-    ) -> Result<ExtRepository, ReductError> {
-        let mut extension_map = IoExtMap::new();
-
-        let query_map = AsyncRwLock::new(HashMap::new());
-        let mut ext_wrappers = Vec::new();
-
-        for path in paths {
-            if !path.exists() {
-                return Err(internal_server_error!(
-                    "Extension directory {:?} does not exist",
-                    path
-                ));
-            }
-
-            for entry in path.read_dir()? {
-                let path = entry?.path();
-                if path.is_file()
-                    && path
-                        .extension()
-                        .map_or(false, |ext| ext == "so" || ext == "dll" || ext == "dylib")
-                {
-                    let ext_wrapper = unsafe {
-                        match Container::<ExtensionApi>::load(path.clone()) {
-                            Ok(wrapper) => wrapper,
-                            Err(e) => {
-                                error!("Failed to load extension '{:?}': {:?}", path, e);
-                                continue;
-                            }
-                        }
-                    };
-
-                    let ext = unsafe { Box::from_raw(ext_wrapper.get_ext(settings.clone())) };
-
-                    info!("Load extension: {:?}", ext.info());
-
-                    let name = ext.info().name().to_string();
-                    extension_map.insert(name, Arc::new(AsyncRwLock::new(ext)));
-                    ext_wrappers.push(ext_wrapper);
-                }
-            }
-        }
-
-        Ok(ExtRepository {
-            extension_map,
-            query_map,
-            ext_wrappers,
-            embedded_extensions,
-        })
-    }
 }
 
 #[async_trait]
@@ -322,70 +267,13 @@ impl ManageExtensions for ExtRepository {
     }
 }
 
-pub fn create_ext_repository(
-    external_path: Option<PathBuf>,
-    embedded_extensions: Vec<Box<dyn ManageStaticAsset + Sync + Send>>,
-    settings: ExtSettings,
-) -> Result<BoxedManageExtensions, ReductError> {
-    if external_path.is_some() || !embedded_extensions.is_empty() {
-        let mut paths = if let Some(path) = external_path {
-            vec![path]
-        } else {
-            Vec::new()
-        };
-
-        for embedded in &embedded_extensions {
-            if let Ok(path) = embedded.absolut_path("") {
-                paths.push(path);
-            }
-        }
-
-        Ok(Box::new(ExtRepository::try_load(
-            paths,
-            embedded_extensions,
-            settings,
-        )?))
-    } else {
-        // Dummy extension repository if
-        struct NoExtRepository;
-
-        #[async_trait]
-        impl ManageExtensions for NoExtRepository {
-            async fn register_query(
-                &self,
-                _query_id: u64,
-                _bucket_name: &str,
-                _entry_name: &str,
-                _query: QueryEntry,
-            ) -> Result<(), ReductError> {
-                Ok(())
-            }
-
-            async fn fetch_and_process_record(
-                &self,
-                _query_id: u64,
-                query_rx: Arc<AsyncRwLock<QueryRx>>,
-            ) -> Option<Result<BoxedReadRecord, ReductError>> {
-                query_rx
-                    .write()
-                    .await
-                    .recv()
-                    .await
-                    .map(|record| record.map(|r| Box::new(r) as BoxedReadRecord))
-            }
-        }
-
-        Ok(Box::new(NoExtRepository))
-    }
-}
+pub(crate) use create::create_ext_repository;
 
 #[cfg(test)]
 pub(super) mod tests {
     use super::*;
     use futures_util::Stream;
     use reduct_base::ext::{Commiter, IoExtensionInfo, Processor};
-    use reqwest::blocking::get;
-    use reqwest::StatusCode;
     use rstest::{fixture, rstest};
 
     use crate::storage::entry::RecordReader;
@@ -397,103 +285,9 @@ pub(super) mod tests {
     use reduct_base::msg::server_api::ServerInfo;
     use reduct_base::Labels;
     use serde_json::json;
-    use std::fs;
     use std::task::{Context, Poll};
     use tempfile::tempdir;
-    use test_log::test as log_test;
 
-    mod load {
-        use super::*;
-        use reduct_base::msg::server_api::ServerInfo;
-        #[log_test(rstest)]
-        fn test_load_extension(ext_repo: ExtRepository) {
-            assert_eq!(ext_repo.extension_map.len(), 1);
-            let ext = ext_repo
-                .extension_map
-                .get("test-ext")
-                .unwrap()
-                .blocking_read();
-            let info = ext.info().clone();
-            assert_eq!(
-                info,
-                IoExtensionInfo::builder()
-                    .name("test-ext")
-                    .version("0.1.1")
-                    .build()
-            );
-        }
-
-        #[log_test(rstest)]
-        fn test_failed_load(ext_settings: ExtSettings) {
-            let path = tempdir().unwrap().keep();
-            fs::create_dir_all(&path).unwrap();
-            fs::write(&path.join("libtest.so"), b"test").unwrap();
-            let ext_repo = ExtRepository::try_load(vec![path], vec![], ext_settings).unwrap();
-            assert_eq!(ext_repo.extension_map.len(), 0);
-        }
-
-        #[log_test(rstest)]
-        fn test_failed_open_dir(ext_settings: ExtSettings) {
-            let path = PathBuf::from("non_existing_dir");
-            let ext_repo = ExtRepository::try_load(vec![path], vec![], ext_settings);
-            assert_eq!(
-                ext_repo.err().unwrap(),
-                internal_server_error!("Extension directory \"non_existing_dir\" does not exist")
-            );
-        }
-
-        #[fixture]
-        fn ext_settings() -> ExtSettings {
-            ExtSettings::builder()
-                .server_info(ServerInfo::default())
-                .build()
-        }
-
-        #[fixture]
-        fn ext_repo(ext_settings: ExtSettings) -> ExtRepository {
-            // This is the path to the build directory of the extension from ext_stub crate
-            const EXTENSION_VERSION: &str = "0.1.1";
-
-            if !cfg!(target_arch = "x86_64") {
-                panic!("Unsupported architecture");
-            }
-
-            let file_name = if cfg!(target_os = "linux") {
-                // This is the path to the build directory of the extension from ext_stub crate
-                "libtest_ext-x86_64-unknown-linux-gnu.so"
-            } else if cfg!(target_os = "macos") {
-                "libtest_ext-x86_64-apple-darwin.dylib"
-            } else if cfg!(target_os = "windows") {
-                "libtest_ext-x86_64-pc-windows-gnu.dll"
-            } else {
-                panic!("Unsupported platform")
-            };
-
-            let ext_path = PathBuf::from(tempdir().unwrap().keep()).join("ext");
-            fs::create_dir_all(ext_path.clone()).unwrap();
-
-            let link = format!(
-                "https://github.com/reductstore/test-ext/releases/download/v{}/{}",
-                EXTENSION_VERSION, file_name
-            );
-
-            let mut resp = get(link).expect("Failed to download extension");
-            if resp.status() != StatusCode::OK {
-                if resp.status() == StatusCode::FOUND {
-                    resp = get(resp.headers().get("location").unwrap().to_str().unwrap())
-                        .expect("Failed to download extension");
-                } else {
-                    panic!("Failed to download extension: {}", resp.status());
-                }
-            }
-
-            fs::write(ext_path.join(file_name), resp.bytes().unwrap())
-                .expect("Failed to write extension");
-
-            let empty_ext_path = tempdir().unwrap().keep();
-            ExtRepository::try_load(vec![ext_path, empty_ext_path], vec![], ext_settings).unwrap()
-        }
-    }
     mod register_query {
         use super::*;
 
@@ -627,16 +421,26 @@ pub(super) mod tests {
         }
 
         #[rstest]
+        #[case(json!({"test-ext": {}, "test-ext2": {}}),  unprocessable_entity!("Multiple extensions are not supported in query id=1"))]
+        #[case(json!({"unknown-ext": {}}),  unprocessable_entity!("Unknown extension 'unknown-ext' in query id=1"))]
+        #[case(json!({}),  unprocessable_entity!("Extension name is not found in query id=1"))]
         #[tokio::test]
-        async fn error_unknown_extension(mut mock_ext: MockIoExtension) {
+        async fn test_error_handling(
+            mut mock_ext: MockIoExtension,
+            processor: BoxedProcessor,
+            commiter: BoxedCommiter,
+            #[case] ext_params: serde_json::Value,
+            #[case] expected_error: ReductError,
+        ) {
             let query = QueryEntry {
-                ext: Some(json!({
-                    "unknown-ext": {},
-                })),
+                ext: Some(ext_params),
                 ..Default::default()
             };
 
-            mock_ext.expect_query().never();
+            mock_ext
+                .expect_query()
+                .with(eq("bucket"), eq("entry"), eq(query.clone()))
+                .return_once(|_, _, _| Ok((processor, commiter)));
 
             let mocked_ext_repo = mocked_ext_repo("test-ext", mock_ext);
             assert_eq!(
@@ -645,7 +449,7 @@ pub(super) mod tests {
                     .await
                     .err()
                     .unwrap(),
-                unprocessable_entity!("Unknown extension 'unknown-ext' in query id=1")
+                expected_error
             );
         }
     }
